reorganized so it's clearer
