+++ requirements.txt
@@ -2,3 +2,4 @@ python-dotenv==1.0.0
 requests==2.31.0
 typing_extensions==4.8.0
 virtualenv==20.25.0
+rich==13.7.0
